Move pool cluster promise test to builtin-runner
